+++ .project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>yoko-leaning</name>
+	<name>yoko-learning</name>
 	<comment></comment>
 	<projects>
 	</projects>
