+++ src/course/the1st/Task1.java
@@ -1,7 +1,7 @@
 package course.the1st;
 
 public class Task1 {
-	
+
 	public static void main(String[] args) {
 		System.out.println("HelloWorld");
 	}
