+++ src/course/the2nd/Task1.java
@@ -1,7 +1,7 @@
 package course.the2nd;
 
 public class Task1 {
-	
+
 	public static void main(String[] args) {
 		String greeting = "HelloWorld";
 		System.out.println(greeting);
