+++ src/course/the2nd/Task2.java
@@ -3,12 +3,12 @@ package course.the2nd;
 public class Task2 {
 	public static void main(String[] args) {
 		int sum = 0;
-		for(int i = 1; i <= 10; i++) {
-			
-			sum += i;	
+		for (int i = 1; i <= 10; i++) {
+
+			sum += i;
 		}
-			System.out.println("１から10の合計は" + sum);
-		
+		System.out.println("１から10の合計は" + sum);
+
 	}
 
 }
